requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-cloud-firestore==2.14.0
google-cloud-aiplatform==1.71.1
anthropic==0.18.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.26.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
gunicorn==21.2.0
google-genai>=0.8.0
fastapi
uvicorn